backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
pdf2docx==0.5.6
python-docx==0.8.11
PyPDF2==3.0.1
Pillow==10.0.1
pdfkit==1.0.0
pandas==2.1.1
pdf2image==1.16.3
openpyxl==3.1.2